Added TS alias.
We need to be able to refer to the current toolset generically
so that loading code does not have to know what toolset is being
loaded. In other words, this code should work:

  module Rubinius::ToolSet.current::TS
  end

Since Ruby forbids (for no good reason) syntax like

  module Rubinius::ToolSet.current
  end

we have to use some concrete module name. So we create an alias.
In other words, the following are equivalent

  Rubinius::ToolSet.current
  Rubinius::ToolSet.current::TS

but one is allowed syntax.